requirements.txt
fastapi
uvicorn
pymongo
openai
apscheduler
requests
python-dotenv
jinja2>=3.0.0
Pillow>=10.0.0
elasticsearch>=8.0.0
httpx==0.24.1  # For async HTTP requests
click==8.1.3  # For CLI